Add poster from Engineering@Southampton Showcase
This poster was displayed at the Engineering@Southampton Showcase event
on 9th June 2015. The print-out has been kept in the broom cupboard. It
could be improved by complementing the University of Southampton logo
with other sponsor logos or removing it.

Images were acquired from Flickr, and used with permission of their
owners (where necessary).

The Eurostile font must be installed to render the poster.